add reset() test
